be/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
httpx==0.27.2
python-jose[cryptography]==3.3.0
